setup_init: skip chrome setup if chrome is installed

+++ setup_init.sh
@@ -40,7 +40,9 @@ fi
 sudo apt purge -y chromium-browser
 
 # install Google Chrome
-wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
-sudo dpkg -i google-chrome-stable_current_amd64.deb
-sudo apt install -f -y
-rm google-chrome-stable_current_amd64.deb
+if ! which google-chrome > /dev/null ; then
+	wget https://dl.google.com/linux/direct/google-chrome-stable_current_amd64.deb
+	sudo dpkg -i google-chrome-stable_current_amd64.deb
+	sudo apt install -f -y
+	rm google-chrome-stable_current_amd64.deb
+fi
